introduce Subject replacing CompatibilitySubject and RegisterSubject
