+++ src/main/kotlin/com/github/imflog/schema/registry/tasks/register/RegisterSubjectExtension.kt
@@ -1,11 +1,8 @@
 package com.github.imflog.schema.registry.tasks.register
 
-import com.github.imflog.schema.registry.LocalReference
-import com.github.imflog.schema.registry.MixedReferenceException
-import com.github.imflog.schema.registry.SchemaType
+import com.github.imflog.schema.registry.Subject
 import com.github.imflog.schema.registry.toSchemaType
 import io.confluent.kafka.schemaregistry.avro.AvroSchema
-import io.confluent.kafka.schemaregistry.client.rest.entities.SchemaReference
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.ListProperty
 
@@ -14,7 +11,7 @@ open class RegisterSubjectExtension(objects: ObjectFactory) {
         const val EXTENSION_NAME = "register"
     }
 
-    val subjects: ListProperty<RegisterSubject> = objects.listProperty(RegisterSubject::class.java)
+    val subjects: ListProperty<Subject> = objects.listProperty(Subject::class.java)
 
     fun subject(inputSubject: String, file: String) = subject(inputSubject, file, AvroSchema.TYPE)
 
@@ -22,27 +19,11 @@ open class RegisterSubjectExtension(objects: ObjectFactory) {
         inputSubject: String,
         file: String,
         type: String
-    ): RegisterSubject {
-        val subject = RegisterSubject(inputSubject, file, type.toSchemaType())
+    ): Subject {
+        val subject = Subject(inputSubject, file, type.toSchemaType())
         subjects.add(subject)
         return subject
     }
-}
 
-data class RegisterSubject(
-    val inputSubject: String,
-    val file: String,
-    val type: SchemaType,
-    val references: MutableList<SchemaReference> = mutableListOf(),
-    val localReferences: MutableList<LocalReference> = mutableListOf()
-) {
-    fun addReference(name: String, subject: String, version: Int): RegisterSubject {
-        references.add(SchemaReference(name, subject, version))
-        return this
-    }
-
-    fun addLocalReference(name: String, path: String): RegisterSubject {
-        localReferences.add(LocalReference(name, path))
-        return this
-    }
+    fun subject(subject: Subject) = subject.apply { subjects.add(subject) }
 }
