+++ src/test/kotlin/com/github/imflog/schema/registry/tasks/register/RegisterTaskActionTest.kt
@@ -1,6 +1,7 @@
 package com.github.imflog.schema.registry.tasks.register
 
 import com.github.imflog.schema.registry.SchemaType
+import com.github.imflog.schema.registry.Subject
 import io.confluent.kafka.schemaregistry.avro.AvroSchema
 import io.confluent.kafka.schemaregistry.avro.AvroSchemaProvider
 import io.confluent.kafka.schemaregistry.client.MockSchemaRegistryClient
@@ -47,7 +48,7 @@ class RegisterTaskActionTest {
         )
 
         val subjects = listOf(
-            RegisterSubject("test", "src/main/avro/external/test.avsc", SchemaType.AVRO)
+            Subject("test", "src/main/avro/external/test.avsc", SchemaType.AVRO)
         )
 
         // when
@@ -91,7 +92,7 @@ class RegisterTaskActionTest {
         )
 
         val subjects = listOf(
-            RegisterSubject(
+            Subject(
                 "test",
                 "src/main/avro/external/test.avsc",
                 SchemaType.AVRO
@@ -174,7 +175,7 @@ class RegisterTaskActionTest {
 
 
         val subjects = listOf(
-            RegisterSubject(
+            Subject(
                 "test",
                 "src/main/avro/external/test.avsc",
                 SchemaType.AVRO
@@ -229,8 +230,8 @@ class RegisterTaskActionTest {
             )
 
         val subjects = listOf(
-            RegisterSubject("test", "src/main/avro/external/test.avsc", SchemaType.AVRO),
-            RegisterSubject("test_2", "src/main/avro/external/test_2.avsc", SchemaType.AVRO),
+            Subject("test", "src/main/avro/external/test.avsc", SchemaType.AVRO),
+            Subject("test_2", "src/main/avro/external/test_2.avsc", SchemaType.AVRO),
         )
 
         // when
